# Conflicts:
#	features/user/otp/src/main/java/com/tokopedia/otp/verification/view/fragment/miscall/MisscallVerificationFragment.kt

+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/inbox/buyerreview/view/adapter/ReputationAdapter.java
@@ -38,9 +38,9 @@ public class ReputationAdapter extends RecyclerView.Adapter<ReputationAdapter.Vi
     public interface ReputationListener {
         void onReputationSmileyClicked(String name, String value);
 
-        void onGoToShopDetail(int shopId);
+        void onGoToShopDetail(long shopId);
 
-        void onGoToPeopleProfile(int userId);
+        void onGoToPeopleProfile(long userId);
     }
 
     ArrayList<SmileyModel> list;
